Fix many problems with publish and install
When trying to install and publish to the remote registry, various
errors were popping up:
- invalid url when the url was passed without the leading http
- unknown library when the path to the remote registry was passed before
the library name: ppm “--registry URL libname”

In the wake of these bugs, the publish, install, and search commands were fixed.
The fix resulted in changing the Get methods in Tokens to Pop so that
the received token is removed from the collection.

Also, the hardcode registry in the auth command was fixed. The registry can now
be passed as an option, or it will be retrieved from the configuration file.